billing_invoice/backend/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
Django
djangorestframework
django-cors-headers
